deepteam_demo/requirements.txt
# DeepTeam Demo Suite Requirements

# Core DeepTeam Framework
deepteam>=1.0.3
deepeval>=3.6.2

# LLM Providers
anthropic>=0.71.0
openai>=1.76.2

# Observability (OpenTelemetry - logging only, no exporters)
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# Data Processing
pandas>=2.0.0

# Console Output
rich>=13.0.0

# CLI
typer>=0.9.0
click>=8.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.2

# Utilities
tqdm>=4.66.1

